NMS-12410 Topology Provider for CDP,LLDP,OSPF,ISIS,Bridge Topologies
Started namespace migration
the classes should use the graph namespace
