+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.linkd/src/test/java/org/opennms/features/topology/plugins/topo/linkd/internal/LinkdTopologyProviderTestIT.java
@@ -52,6 +52,7 @@ import org.opennms.features.topology.api.topo.Edge;
 import org.opennms.features.topology.api.topo.EdgeRef;
 import org.opennms.features.topology.api.topo.Vertex;
 import org.opennms.features.topology.api.topo.VertexRef;
+import org.opennms.features.topology.api.topo.DefaultTopologyProviderInfo;
 import org.opennms.netmgt.dao.api.GenericPersistenceAccessor;
 import org.opennms.netmgt.enlinkd.BridgeOnmsTopologyUpdater;
 import org.opennms.netmgt.enlinkd.CdpOnmsTopologyUpdater;
@@ -86,7 +87,7 @@ public class LinkdTopologyProviderTestIT {
     GenericPersistenceAccessor genericPersistenceAccessor;
 
     @Autowired
-    LinkdTopologyProvider linkdTopologyProvider;
+    LinkdTopologyFactory linkdTopologyFactory;
 
     @Autowired
     TopologyEntityCache entityCache;
@@ -117,6 +118,8 @@ public class LinkdTopologyProviderTestIT {
 
     private TopologyGenerator generator;
 
+    private LinkdTopologyProvider linkdTopologyProvider;
+
     @BeforeTransaction
     public void setUp() {
         nodesOnmsTopologyUpdater.register();
@@ -132,6 +135,9 @@ public class LinkdTopologyProviderTestIT {
         generator = TopologyGenerator.builder()
                 .persister(persister)
                 .progressCallback(progressCallback).build();
+
+        linkdTopologyProvider = new LinkdTopologyProvider(linkdTopologyFactory);
+        linkdTopologyProvider.setTopologyProviderInfo(new DefaultTopologyProviderInfo());
     }
 
     @Test
