+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.linkd/src/test/resources/META-INF/opennms/applicationContext-LinkdTopologyProviderTestIT.xml
@@ -207,7 +207,7 @@
         <constructor-arg ref="nodeTopologyService" />
     </bean>
 
-    <bean id="enLinkdTopologyProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.LinkdTopologyProvider">
+    <bean id="enLinkdTopologyProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.LinkdTopologyFactory">
         <constructor-arg ref="metricRegistry"/>
         <property name="onmsTopologyDao" ref="onmsTopologyDao" />
     </bean>
